[API] Implement payment method group generator
