+++ features/admin/payment/managing_payment_methods/adding_payment_method.feature
@@ -8,7 +8,7 @@ Feature: Adding a new payment method
         Given the store operates on a single channel in "United States"
         And I am logged in as an administrator
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new payment method
         When I want to create a new offline payment method
         And I name it "Offline" in "English (United States)"
@@ -17,7 +17,7 @@ Feature: Adding a new payment method
         Then I should be notified that it has been successfully created
         And the payment method "Offline" should appear in the registry
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new payment method with description
         When I want to create a new offline payment method
         And I name it "Offline" in "English (United States)"
@@ -27,7 +27,7 @@ Feature: Adding a new payment method
         Then I should be notified that it has been successfully created
         And the payment method "Offline" should appear in the registry
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new payment method with instructions
         When I want to create a new offline payment method
         And I name it "Offline" in "English (United States)"
@@ -38,7 +38,7 @@ Feature: Adding a new payment method
         And the payment method "Offline" should appear in the registry
         And the payment method "Offline" should have instructions "Bank account: 0000 1111 2222 3333" in "English (United States)"
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new payment method for channel
         When I want to create a new offline payment method
         And I name it "Offline" in "English (United States)"
@@ -49,7 +49,7 @@ Feature: Adding a new payment method
         And the payment method "Offline" should appear in the registry
         And the payment method "Offline" should be available in channel "United States"
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new paypal payment method
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I name it "Paypal Express Checkout" in "English (United States)"
@@ -59,7 +59,7 @@ Feature: Adding a new payment method
         Then I should be notified that it has been successfully created
         And the payment method "Paypal Express Checkout" should appear in the registry
 
-    @todo-api @ui
+    @api @ui
     Scenario: Adding a new stripe payment method
         When I want to create a new payment method with "Stripe Checkout" gateway factory
         And I name it "Stripe Checkout" in "English (United States)"
