+++ features/admin/payment/managing_payment_methods/payment_method_unique_code_validation.feature
@@ -9,7 +9,7 @@ Feature: Payment method unique code validation
         And the store has a payment method "Offline" with a code "Offline"
         And I am logged in as an administrator
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add payment method with taken code
         When I want to create a new offline payment method
         And I name it "Paypal Express Checkout" in "English (United States)"
