+++ features/admin/payment/managing_payment_methods/payment_method_validation.feature
@@ -18,14 +18,14 @@ Feature: Payment method validation
         Then I should be notified that code is required
         And the payment method with name "Paypal Express Checkout" should not be added
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new payment method translation in unexisting language
         When I want to modify the "Offline" payment method
         And I name it "Offline" in "French (France)"
         And I try to save my changes
         Then I should be notified that the locale is not available
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add a new payment method with a too long code
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I name it "Paypal Express Checkout" in "English (United States)"
@@ -33,7 +33,7 @@ Feature: Payment method validation
         And I add it
         Then I should be notified that code is too long
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add a new payment method without specifying its name
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I specify its code as "PEC"
@@ -42,7 +42,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify payment method name
         And the payment method with code "PEC" should not be added
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add a new paypal payment method without specifying password
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I name it "Paypal Express Checkout" in "English (United States)"
@@ -53,7 +53,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify paypal password
         And the payment method with code "PEC" should not be added
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new paypal payment method without specifying sandbox
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I name it "Paypal Express Checkout" in "English (United States)"
@@ -63,7 +63,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify paypal sandbox status
         And the payment method with code "PEC" should not be added
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new paypal payment method, but with sandbox that has wrong type
         When I want to create a new payment method with "Paypal Express Checkout" gateway factory
         And I name it "Paypal Express Checkout" in "English (United States)"
@@ -73,7 +73,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify paypal sandbox status that is boolean
         And the payment method with code "PEC" should not be added
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add a new stripe payment method with only publishable key specified
         When I want to create a new payment method with "Stripe Checkout" gateway factory
         And I name it "Stripe Checkout" in "English (United States)"
@@ -83,7 +83,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify stripe "Secret key"
         And the payment method with code "PEC" should not be added
 
-    @todo-api @ui
+    @api @ui
     Scenario: Trying to add a new stripe payment method with only secret key specified
         When I want to create a new payment method with "Stripe Checkout" gateway factory
         And I name it "Stripe Checkout" in "English (United States)"
@@ -93,7 +93,7 @@ Feature: Payment method validation
         Then I should be notified that I have to specify stripe "Publishable key"
         And the payment method with code "PEC" should not be added
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new payment method without gateway configuration
         When I want to create a new payment method without gateway configuration
         And I name it "Payment method without gateway configuration" in "English (United States)"
@@ -101,7 +101,7 @@ Feature: Payment method validation
         And I try to add it
         Then I should be notified that I have to specify gateway configuration
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new payment method without gateway name
         When I want to create a new payment method without gateway name
         And I name it "Payment method without gateway name" in "English (United States)"
@@ -109,7 +109,7 @@ Feature: Payment method validation
         And I try to add it
         Then I should be notified that I have to specify gateway name
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new payment method without factory name
         When I want to create a new payment method without factory name
         And I name it "Payment method without gateway factory name" in "English (United States)"
@@ -117,7 +117,7 @@ Feature: Payment method validation
         And I try to add it
         Then I should be notified that I have to specify factory name
 
-    @todo-api @no-ui
+    @api @no-ui
     Scenario: Trying to add a new payment method with wrong factory name
         When I want to create a new payment method with wrong factory name
         And I name it "Payment method with wrong gateway factory name" in "English (United States)"
