+++ features/shop/checkout/paying_for_order/preventing_to_pay_for_cancelled_order.feature
@@ -13,7 +13,7 @@ Feature: Preventing to pay for the cancelled order
         And the store allows paying Offline
         And there is a customer "sylius@example.com" that placed an order "#00000022"
 
-    @todo-api @ui
+    @api @ui
     Scenario: Not being able to pay for cancelled order
         Given the customer bought 3 "Iron Maiden T-Shirt" products
         And the customer chose "Free" shipping method to "United States" with "Offline" payment
